requirements.txt
pandas
numpy
requests
Pillow
tqdm
torch
transformers>=4.30
scikit-learn
